+++ src/servers/mod.rs
@@ -6,10 +6,6 @@ pub mod rapier_physics_server_2d;
 #[cfg(feature = "dim3")]
 pub mod rapier_physics_server_3d;
 pub mod rapier_physics_server_extra;
-#[cfg(feature = "dim2")]
-pub mod rapier_physics_server_factory_2d;
-#[cfg(feature = "dim3")]
-pub mod rapier_physics_server_factory_3d;
 pub mod rapier_physics_server_impl;
 pub mod rapier_project_settings;
 #[cfg(feature = "dim2")]
@@ -21,15 +17,14 @@ pub fn register_server() {
     use godot::engine::PhysicsServer2DManager;
     let mut manager = PhysicsServer2DManager::singleton();
     let factory =
-        crate::servers::rapier_physics_server_factory_2d::RapierPhysicsServerFactory2D::new_alloc();
+        crate::servers::rapier_physics_server_2d::RapierPhysicsServerFactory2D::new_alloc();
     manager.register_server("Rapier2D".into(), factory.callable("create_server"));
 }
 #[cfg(feature = "dim3")]
 pub fn register_server() {
     use godot::engine::PhysicsServer3DManager;
     let mut manager = PhysicsServer3DManager::singleton();
-    let factory =
-        crate::servers::rapier_physics_server_factory_3d::RapierPhysicsServerFactory3D::new_alloc();
+    let factory = crate::servers::rapier_physics_3d::RapierPhysicsServerFactory3D::new_alloc();
     manager.register_server("Rapier3D".into(), factory.callable("create_server"));
 }
 pub fn register_scene() {
